+++ test/test_snapshot.py
@@ -95,6 +95,7 @@ class Snapshot:
 
 
 snapshots = [
+    Snapshot("ast_parsing"),
     Snapshot("demo", min_version=(3, 9)),
     Snapshot("flavors_google"),
     Snapshot("flavors_numpy"),