+++ test/testdata/ast_parsing.py
@@ -1,0 +1,5 @@
+class Foo:
+    def __init__(self):
+        self.no_docstring = 42
+        self.with_docstring = 43
+        """This is an attribute docstring."""