+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 <!-- ✨ You do not need to add a pull request reference or an author, this will be added automatically by CI. ✨ -->
 
+ - pdoc now documents variables by default, even if they have no docstring or type annotation.
+   Please make yourself heard in [#574](https://github.com/mitmproxy/pdoc/issues/574) if that
+   introduces significant issues for your use case.
+   ([#575](https://github.com/mitmproxy/pdoc/pull/575), @mhils)
  - Add support for Python 3.12.
    ([#570](https://github.com/mitmproxy/pdoc/pull/570), @mhils)
  - Remove support for Python 3.7, which has reached end-of-life on 2023-06-27.